Clean up setting component state, split it from configuration
Introduce shelly::Component::SetState() that is similar to ::SetConfig() but is for runtime state (like setting state of the switch or opening garage door).

Previosuly we were sneaking in state changes as part of config and that's not nice.

RPC API changes as well: Shelly.SetSwitch is gone, instead we have a generic Shelly.SetState that takes a component id and type and opaque JSON state that is passed to component for processing.

In practical terms,

```
$ curl http://shelly1-test.local/rpc/Shelly.SetSwitch -d '{"id": 1, "state": true}'
```
becomes
```
$ curl http://shelly1-test.local/rpc/Shelly.SetState -d '{"id": 1, "type": 0, "state": {"state": true}}'
```
or in GET format:
```
$ curl http://shelly1-test.local/rpc/Shelly.SetState?id=1&type=0&state=%7b%22state%22%3atrue%7d
```