Improve failsafe mode
Some devices have button connected to a strap pin and cannot be booted with button down (2.5, 1PM).
Instead, we modify the algorithm to allow up to 3 seconds on power up for user to press and hold the button for at least 2 seconds.
During wait LED is blinking fast (100 ms on/off) and stops blinking when
the button is pressed. It then lights up solid when device boots into failsafe mode.

Soft reboots do not have this delay, so most of reboots are just as fast.