compatibilty with upgrade from 1.4.0
1.4.0 enlarges the app partition by 0x10000 bytes (64KB), taking it from
  fs

the new firmware (1.4.x) checks the new manifest for the new file system size. if it
does not match it says "PT Error".

Also core is upgraded to include newest lfs (2.1 filesystem version).
Otherwise upgrade fails as the old filesystem cannot be migrated.

This is still work in progress; we cannot revert to 1.4.0 yet.